// src/app/types/ip-addr.ts
export type IpAddr = {
    query: string;
    status: string;
    continent: string;
    continent_code: string;
    country: string;
    country_code: string;
    region: string;
    region_name: string;
    city: string;
    district: string;
    zip_code: string;
    lat: number;
    lon: number;
    timezone: string;
    offset: number;
    isp: string;
    organization: string;
    as_number: string;
    as_name: string;
    reverse_dns: string;
    is_mobile: boolean;
    is_proxy: boolean;
    is_hosting: boolean;
  }
  